refactor: compute worldTransform of 360 entities in factory
